refactor(desktop): fold the renderer one-caller files
