+++ apps/desktop/src/renderer/app.tsx
@@ -104,7 +104,6 @@ import { encodeFeedbackImage } from "./feedback-images";
 import { FeedbackSlot } from "./feedback-slot";
 import { KeySlot } from "./key-slot";
 import { type Errand, errandTargets, LukeErrand } from "./luke-errand";
-import { usePrefersReducedMotion } from "./luke-face-mood";
 import { applySpokenSetting, buildLukeGuide, isAppSettingId } from "./luke-guide";
 import { MarkdownMessage } from "./markdown-message";
 import { NotchWings } from "./notch-wings";
@@ -163,6 +162,7 @@ import { useBootstrapRacedChannel } from "./use-bootstrap-raced-channel";
 import { useMeasuredHeight } from "./use-measured-height";
 import { panelEntryOpen, usePanelEntry } from "./use-panel-entry";
 import { usePanelPresentation } from "./use-panel-presentation";
+import { usePrefersReducedMotion } from "./use-reduced-motion";
 import { useStateWithRef } from "./use-state-with-ref";
 import { useVoiceView, voiceErrorToShow, voiceNoticeToShow } from "./use-voice-view";
 import type { AppActionCarrier } from "./voice/realtime-session";
