+++ apps/desktop/src/renderer/luke-face-mood.ts
@@ -312,24 +312,6 @@ export function useFaceHover(face: RefObject<HTMLElement | null>): boolean {
   return hovered;
 }
 
-/** Reduced motion is a setting, not a media query the stylesheet alone can answer:
- * holding every loop still leaves the poses, and switching between poses is the
- * motion someone asked not to see. */
-export function usePrefersReducedMotion(): boolean {
-  const [reduced, setReduced] = useState(
-    () => window.matchMedia("(prefers-reduced-motion: reduce)").matches,
-  );
-
-  useEffect(() => {
-    const query = window.matchMedia("(prefers-reduced-motion: reduce)");
-    const update = () => setReduced(query.matches);
-    query.addEventListener("change", update);
-    return () => query.removeEventListener("change", update);
-  }, []);
-
-  return reduced;
-}
-
 /**
  * The gesture holding the face, and which play of it this is. The count is what
  * lets the same gesture happen twice running: a motion now plays once rather
