+++ apps/desktop/src/renderer/settings-panel.tsx
@@ -14,8 +14,35 @@ import {
   providerRunsSessionsInCloud,
   VOICE_CREDENTIAL_PROVIDER,
 } from "@sidecar/credentials/vocabulary";
+import { FEEDBACK_KIND, type FeedbackKind } from "@sidecar/feedback";
 import { APP_SETTING_KIND, APP_TOGGLE_VALUE } from "@sidecar/guide";
-import { CloudBadge, ProviderMark } from "@sidecar/panel";
+import {
+  BackIcon,
+  CheckIcon,
+  ChevronIcon,
+  CloseIcon,
+  CloudBadge,
+  CloudIcon,
+  DisplayIcon,
+  DownloadIcon,
+  ExternalIcon,
+  FolderIcon,
+  KeyboardIcon,
+  KeyIcon,
+  LukeIcon,
+  MegaphoneIcon,
+  PencilIcon,
+  PlugIcon,
+  PopUpIcon,
+  PowerIcon,
+  ProviderMark,
+  RefreshIcon,
+  ResetIcon,
+  ShieldIcon,
+  SpeakerIcon,
+  TrashIcon,
+  UserIcon,
+} from "@sidecar/panel";
 import {
   isRealtimeVoice,
   isRealtimeVoiceSpeed,
@@ -102,8 +129,7 @@ import {
   removalWithdrawable,
 } from "./credential-removal";
 import { DestinationNote } from "./destination-note";
-import type { FeedbackEntryControl } from "./feedback-entry";
-import { FeedbackSection } from "./feedback-panel";
+import { FEEDBACK_COPY, type FeedbackEntryControl } from "./feedback-entry";
 import { Keycaps } from "./keycaps";
 import { type ErrandTarget, errandTargetProps } from "./luke-errand";
 import { APP_SETTING_ID } from "./luke-guide";
@@ -117,30 +143,6 @@ import {
   voiceSourceLabel,
 } from "./microphone-access";
 import { PANEL_TAB, panelPanelId, panelTabId } from "./panel-tabs";
-import {
-  BackIcon,
-  CheckIcon,
-  ChevronIcon,
-  CloseIcon,
-  CloudIcon,
-  DisplayIcon,
-  DownloadIcon,
-  ExternalIcon,
-  FolderIcon,
-  KeyboardIcon,
-  KeyIcon,
-  LukeIcon,
-  PencilIcon,
-  PlugIcon,
-  PopUpIcon,
-  PowerIcon,
-  RefreshIcon,
-  ResetIcon,
-  ShieldIcon,
-  SpeakerIcon,
-  TrashIcon,
-  UserIcon,
-} from "./settings-icons";
 import {
   defaultProjectRowId,
   landOnSettingsRow,
@@ -3544,6 +3546,67 @@ function UpdatesSection({
           <ExternalIcon />
         </button>
       </div>
+    </section>
+  );
+}
+
+/**
+ * One button that offers a kind: its name is the whole line, and pressing it
+ * opens the composer on it. No sentence under it — the composer's own shape is
+ * where each kind explains itself. The button says "keep writing" when a draft
+ * of this kind is waiting, because pressing it then continues rather than
+ * opens.
+ */
+function FeedbackOffer({
+  kind,
+  control,
+}: {
+  kind: FeedbackKind;
+  control: FeedbackEntryControl;
+}): React.JSX.Element {
+  const copy = FEEDBACK_COPY[kind];
+  const holdsDraft = control.entry?.kind === kind;
+  return (
+    <button
+      type="button"
+      className="quiet-button feedback-offer"
+      onClick={() => control.begin(kind)}
+    >
+      {holdsDraft ? "Keep writing" : copy.title}
+    </button>
+  );
+}
+
+/**
+ * The two ways to write to the people who make Luke, side by side on one row,
+ * standing on the settings front page above the account and the way out. The
+ * section only offers — pressing either button stands the panel down to the
+ * composer's own shape, the way pressing Connect stands it down to the key
+ * slot — and the line under the offers is where a landed send reports back,
+ * which is why leaving the composer comes back to this page rather than to
+ * wherever it was last.
+ */
+function FeedbackSection({ control }: { control: FeedbackEntryControl }): React.JSX.Element {
+  return (
+    <section
+      className="settings-section"
+      style={cssCustomProperties({ "--row-index": 3 })}
+      {...searchAnchorProps(SETTINGS_SEARCH_ROW.FEEDBACK)}
+    >
+      <h2>
+        <MegaphoneIcon />
+        Feedback
+      </h2>
+      <div className="feedback-offers">
+        <FeedbackOffer kind={FEEDBACK_KIND.FEEDBACK} control={control} />
+        <FeedbackOffer kind={FEEDBACK_KIND.PROMPT} control={control} />
+      </div>
+      {control.notice ? (
+        <p className="feedback-sent">
+          <CheckIcon />
+          {control.notice}
+        </p>
+      ) : null}
     </section>
   );
 }
