+++ apps/desktop/src/renderer/use-reduced-motion.ts
@@ -1,0 +1,19 @@
+import { useEffect, useState } from "react";
+
+/** Reduced motion is a setting, not a media query the stylesheet alone can answer:
+ * holding every loop still leaves the poses, and switching between poses is the
+ * motion someone asked not to see. */
+export function usePrefersReducedMotion(): boolean {
+  const [reduced, setReduced] = useState(
+    () => window.matchMedia("(prefers-reduced-motion: reduce)").matches,
+  );
+
+  useEffect(() => {
+    const query = window.matchMedia("(prefers-reduced-motion: reduce)");
+    const update = () => setReduced(query.matches);
+    query.addEventListener("change", update);
+    return () => query.removeEventListener("change", update);
+  }, []);
+
+  return reduced;
+}
